Add IFort and MKL support to macOS compiles. (#9)
* Updated CMake codes to select between MKL, FFTW3, Accelerate or system BLAS/LAPACK


* Fixes for Windows running IFort

* Fixes crashes at end of GetADPImage and other fixes

* Updates for better stability and compatibility

* Removes old build scripts
* Updates CMake code to use target properties for compile and linker flags
* Quiet warnings in H5ScopedSentinel
* Update include directives for H5ScopedSentinel
* Remove hard coded file paths in MonteCarlo UI
* Fix Fortran compile warnings about already 'used' modules

* Add missing code to enable the UI after the ADP is calculated


* Cleaning up compiler warnings on MSVC


* Change from std::vector to QVector to remove compiler warnings about DLL
interface


* Applications should be GUI Applications


* Update MonteCarlo and MasterPattern to write out NML data sets to HDF5

* Before the code was writing JSON data array which has not been fully implemented
inside of EMsoft at a low level
* Default values for the gui should allow a user to index the sample data set from EDAX
that can be downloaded.


* remove spurious '_' character in Cmake code


* MonteCarlo now uses QProcess to call out to the actual EMMCOpenCL
executable.O


* EMEBSDmaster module now uses a QProcess to execute the EMEBSDmaster
program

NML File is written to a temp location while the program executes
Misc changes in MonteCarlo module.


* Read from std error when trying to get the error messages from EMsoft programs


* Use correct install location for H5Support on macOS


* Implement method to get the path to the EMSOFTPATHNAME variable


* Fix packaging issues where the Style plugin was missing


* Export the EMSOFTPATHNAME env variable for each executable

More explicit output for the executable is logged.


* Various small code cleanups in CMake


* Updates to Workbench to allow file consistency, canceling of sub-process


* Fix packaging errors on macOS due to missing library path for iomp5.dylib


* Fix issues with MKL LAPACK ZGEEV using F77 semantics


* Add unit test for EMEBSD executable.


* Set defaults for generating patterns in the Workbench application


* Add the -heap-arrays compiler argument for IFort compilers.


* Disable warnings about the DLL export comments on Windows

This is really only needed if we use /stand:f03 as an additional argument to the compiler.


* Only use a single thread when generating patterns.

Using more than a single thread causes crashes.


* Use a default value of 0.333 for the Gamma setting.


* Add required C++ header include directives


* Pass by const reference where possible.


* Move GFortran compiler flags to modern cmake syntax.

The flags are now set in the EMsoftLib part of the CMake code as PUBLIC compiler flags
which means anything that links to EMsoftLib will get those arguments.


* Allow the developer to over ride the number of processors used for the unit tests

+++ Source/EMsoftWorkbench/Workbench/Modules/CrystalStructureCreationModule/AsymmetricUnitTableData.cpp
@@ -140,12 +140,12 @@ bool AsymmetricUnitTableData::isEmpty() const
 // -----------------------------------------------------------------------------
 void AsymmetricUnitTableData::checkAndAdjustDimensions()
 {
-  QSize dataSize(m_TableData.size(), 0);
+  QSize dataSize(static_cast<int32_t>(m_TableData.size()), 0);
   QSize headerSize(m_RowHeaders.size(), m_ColHeaders.size());
 
   if(!m_TableData.empty())
   {
-    dataSize.setHeight(m_TableData[0].size());
+    dataSize.setHeight(static_cast<int32_t>(m_TableData[0].size()));
   }
 
   if(dataSize == headerSize)
@@ -489,7 +489,7 @@ void AsymmetricUnitTableData::setTableData(const std::vector<std::vector<double>
 // -----------------------------------------------------------------------------
 int AsymmetricUnitTableData::getNumRows() const
 {
-  return m_TableData.size();
+  return static_cast<int32_t>(m_TableData.size());
 }
 
 // -----------------------------------------------------------------------------
@@ -499,7 +499,7 @@ int AsymmetricUnitTableData::getNumCols() const
 {
   if(!m_TableData.empty())
   {
-    return m_TableData[0].size();
+    return static_cast<int32_t>(m_TableData[0].size());
   }
 
   return 0;
@@ -545,9 +545,9 @@ bool AsymmetricUnitTableData::operator==(const AsymmetricUnitTableData& rhs) con
   if(m_RowHeaders == rhs.m_RowHeaders && m_ColHeaders == rhs.m_ColHeaders && m_DynamicRows == rhs.m_DynamicRows && m_DynamicCols == rhs.m_DynamicCols && m_MinRows == rhs.m_MinRows &&
      m_MinCols == rhs.m_MinCols && m_DefaultRowCount == rhs.m_DefaultRowCount && m_DefaultColCount == rhs.m_DefaultColCount)
   {
-    for(int i = 0; i < m_TableData.size(); i++)
+    for(size_t i = 0; i < m_TableData.size(); i++)
     {
-      for(int j = 0; j < m_TableData[i].size(); j++)
+      for(size_t j = 0; j < m_TableData[i].size(); j++)
       {
         if(m_TableData[i][j] != rhs.m_TableData[i][j])
         {
@@ -569,9 +569,9 @@ bool AsymmetricUnitTableData::operator!=(const AsymmetricUnitTableData& rhs) con
   if(m_RowHeaders == rhs.m_RowHeaders && m_ColHeaders == rhs.m_ColHeaders && m_DynamicRows == rhs.m_DynamicRows && m_DynamicCols == rhs.m_DynamicCols && m_MinRows == rhs.m_MinRows &&
      m_MinCols == rhs.m_MinCols && m_DefaultRowCount == rhs.m_DefaultRowCount && m_DefaultColCount == rhs.m_DefaultColCount)
   {
-    for(int i = 0; i < m_TableData.size(); i++)
+    for(size_t i = 0; i < m_TableData.size(); i++)
     {
-      for(int j = 0; j < m_TableData[i].size(); j++)
+      for(size_t j = 0; j < m_TableData[i].size(); j++)
       {
         if(m_TableData[i][j] != rhs.m_TableData[i][j])
         {
